+++ lib/datarockets/style.rb
@@ -5,6 +5,7 @@ require "datarockets/style/version"
 
 require "datarockets/style/cop/layout/array_alignment_extended"
 require "datarockets/style/cop/style/nested_interpolation"
+require "datarockets/style/cop/rspec/prefer_before_over_setup"
 
 module Datarockets
   # Datarickors sharable config