+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is described in [Contributing notes](CONTRIBUTING.md#changelog-entry-
 
 ## master
 
+* Add `RSpec/PreferBeforeOverSetup` cop. ([@paydaylight](https://github.com/paydaylight))
+
 ## 1.0.1 (2021-02-04)
 
 ### Added